+++ bytecodelearners-vue-js/.env.development
@@ -1,4 +1,4 @@
-VUE_APP_SERVER_URL=http://localhost:8000/api
-VUE_APP_IMAGE_URL=http://localhost:8000/storage/
+VUE_APP_SERVER_URL=https://www.bytecodelearners.club/bytecode-server/api
+VUE_APP_IMAGE_URL=https://www.bytecodelearners.club/bytecode-server/storage/app/public/
 VUE_APP_SERVER_STATIC_FILES=https://www.bytecodelearners.club/statics
-VUE_APP_SERVER_STORAGE_FILES=http://localhost:8000/storage
+VUE_APP_SERVER_STORAGE_FILES=https://www.bytecodelearners.club/bytecode-server/storage/app/
