+++ bytecodelearners-vue-js/src/components/Layout/ByteCodeNavigationDrawer.vue
@@ -16,7 +16,7 @@
         v-if="getCurrentRoute == '/'"
       >
         <v-icon class="bcl-menu-item-icon" color="green darken-2" x-large
-          >mdi-party-popper</v-icon
+          >mdi-calendar-star</v-icon
         >
         <div class="bcl-menu-item-title bcl-large-text">Events</div>
       </div>
@@ -70,6 +70,15 @@
           >mdi-wall</v-icon
         >
         <div class="bcl-menu-item-title bcl-medium-text">Initiator</div>
+      </div>
+      <div
+        class="bcl-menu-item"
+        @click="routerView('/events_winner')"
+      >
+        <v-icon class="bcl-menu-item-icon" color="green darken-2" x-large
+          >mdi-party-popper</v-icon
+        >
+        <div class="bcl-menu-item-title bcl-medium-text">Events Winner</div>
       </div>
       <div
         class="bcl-menu-item"
