+++ bytecodelearners-vue-js/src/views/winners/event1/css/4 style.css
@@ -1,0 +1,86 @@
+div {
+    border:3px solid red;
+    margin:3px;
+    padding:4px;
+    animation:colorchg 5s infinite;
+}
+@keyframes colorchg{
+    0%{
+        border:3px solid green;
+    }
+    25%{
+        border:3px solid yellow;
+    }
+    50%{
+        border:3px solid cyan;
+    }
+    75%{
+        border:3px solid magenta;
+    }
+    100%{
+        border:3px solid blue;
+    }
+}
+.flex{
+    display:flex;
+}
+
+
+
+.div1{
+     <span class="blinking">EXAM</span>
+    border-top-style:double ;
+    border-box: 20%px; 
+    border-width:10px;
+    border-color:black  ;
+    text-top :exam;
+    text-align:exam;
+    box-text:exam;
+     .btn {
+  border: none;
+  padding: 14px 28px;
+  font-size: 16px;
+  cursor: pointer;
+  display: inline-block;
+}
+
+/* On mouse-over */
+.btn:hover {background: #eee;}
+
+
+     
+     
+}
+.div2{
+     
+    border-box: 20%px; 
+    border-width:10px;
+    border-color:black  ;
+    text-top :exam;
+    text-align:exam;
+    text-shadow:black;
+    box-text:exam;
+     
+     
+}
+.div3{
+     
+    border-width:10px;
+    border-color:black  ;
+    text-top :exam;
+    text-align:exam;
+    text-shadow:black;
+    box-text:exam;
+     
+}
+.div4{
+
+    border-box: 20%px; 
+    border-width:10px;
+    border-color:black;
+    text-top :exam;
+    text-align:exam;
+    box-text:exam;
+     
+}
+
