+++ bytecodelearners-vue-js/src/views/winners/event1/css/aboutus.css
@@ -1,0 +1,46 @@
+.body{
+     margin: 0;
+padding: 0;
+font-family: Century Gothic;
+}
+.header{
+    height: 100vh;
+    width: 100%;
+    background-color: rgba(111, 235, 193, 0.795);
+    background-size: cover;
+    background-position: center;
+}
+.main{
+    width: 50%;
+    background-color: rgba(134, 234, 240, 0.459);
+    text-align:center;
+    outline-color:rgb(8, 8, 8);
+    outline-width: thick;
+    word-break:break-word;
+    padding: 0%;
+    position: absolute;
+    top: 10%;
+    left: 50%;
+    transform: translate(-50%);
+    border-top-style:groove
+    
+}
+.btn{
+    padding: 10px 20px;
+    font-family: Century Gothic;
+    font-size: large;
+    font-weight: 500;
+    border-style: solid;
+    outline-width:20px;
+    outline-color:whitesmoke;
+    background-color: white;
+    border-style: ridge;
+    cursor: pointer;
+    transition: 0.3s ease-in-out;
+}
+.btn:hover{
+    background-color: aqua;
+    color:black
+}
+
+
