+++ bytecodelearners-vue-js/src/views/winners/event1/css/box.css
@@ -1,0 +1,142 @@
+.title {
+  text-align: center;
+  font-size: 4.5rem;
+  font-weight: bold;
+ 
+}
+
+.box1 {
+  font-weight: bold;
+  float: left;
+  font-size: 25px;
+  text-align: center;
+  margin-top: 5px;
+  
+
+  
+}
+
+#a {
+  background-color: skyblue;
+  color: rgb(238, 245, 245);
+  width: 23%;
+  height:70%;
+  margin-left: 14px;
+  background-color: rgb(150, 221, 252);
+  
+}
+
+#b {
+  background-color: skyblue;
+  color: rgb(244, 252, 251);
+  width: 23%;
+  height: 70%;
+  border: 0px solid skyblue;
+  margin-left: 14px;
+  
+}
+
+#c {
+  background-color: skyblue;
+  color: rgb(250, 253, 253);
+  width: 23%;
+  height: 70%;
+  border: 0px solid skyblue;
+  margin-left: 14px;
+  
+}
+
+#d {
+  background-color: skyblue;
+  color: white;
+  width: 25%;
+  height: 70%;
+  border: 0px solid skyblue;
+  margin-left: 14px;
+ 
+}
+
+#g {
+  list-style: none;
+  color: black;
+  font-size: 35px;
+  margin-left: -50px;
+  background-color: white;
+  
+}
+#m {
+  background-color: rgba(126, 245, 229, 0.651);
+}
+
+#k {
+  background-color: white;
+  color: black;
+  font-size: 20px;
+  font-weight: bold;
+  border-color: white;
+}
+
+@media screen and (max-width:1200px){
+.box1{
+  margin: 40px 30px 0px 40px;
+  height: 20%;
+  overflow:scroll;}
+  #a{
+    width: 100%;
+  height: 430px;}
+  #b{
+  width: 100%;
+  height: 430px; }
+  #c{
+    width: 100%;
+    height: 430px; } 
+  #d{
+     width: 100%;
+     height: 430px; }
+
+
+}
+
+
+
+
+/* @media only screen and (max-width: 350px) and (min-width:100px  )  {
+    .title{
+        text-align: center;font-size: 1.5rem;
+    }
+} */
+/* 
+@media only screen and (max-width: 640px) {
+  #a {
+    width: 100%;
+    height: 100%;
+    margin-top: 2px;
+  }
+
+  #b {
+    width: 100%;
+    height: 100%;
+    margin-top: 2px;
+  }
+  #c {
+    width: 100%;
+    height: 100%;
+     margin-top: 2px;
+
+  } */
+
+  /* #d {
+    width: 100%;
+    height: 100%;
+    margin-top: 2px;
+  }
+
+  #g {
+    color: rgb(168, 22, 22);
+    background-color: chartreuse;
+
+  }
+  #m {
+    background-color: crimson;
+  }
+} */
