+++ bytecodelearners-vue-js/src/views/winners/event1/css/boxstyl.css
@@ -1,0 +1,99 @@
+p {
+    text-align: center;
+  }
+  div {
+    text-align: center;
+  }
+
+div {
+    border:30px solid red;
+    margin: 30px;
+    padding:40px;
+    animation:colorchg 20s infinite;
+}
+.div1 {
+    word-wrap: break-word;
+  }
+  .div2 {
+    word-wrap: break-word;
+  }
+  .div3 {
+    word-wrap: break-word;
+  }
+  .div4 {
+    word-wrap: break-word;
+  }
+@keyframes colorchg{
+    0%{
+        border:10px solid green;
+    }
+    25%{
+        border:10px solid yellow;
+    }
+    50%{
+        border:10px solid cyan;
+    }
+    75%{
+        border:10px solid magenta;
+    }
+    100%{
+        border:10px solid blue;
+    }
+}
+.flex{
+    display:flex;
+}
+
+
+
+ .div1{
+    height:22%;
+    width: 22%;
+    background-color: teal;
+    border-width:10px;
+    
+     
+     
+}
+.div2{
+     
+        height: 22%;
+        width: 22%;
+        background-color: teal;
+        position: relative;
+       
+    
+    
+     
+     
+}
+.div3{
+    height: 22%;
+    width: 22%;
+    background-color: teal;
+    position: relative;
+    border-color:black  ;
+  
+  
+     
+}
+.div4{
+    height: 22%;
+    width: 22%;
+    background-color: teal;
+    position: relative;
+    border-width:10px;
+    border-color:black;
+   
+   
+     
+}
+@media only screen and (max-width:600px) and(min-width:360){
+    .div1{
+        height:100%;
+        width: 100%;
+        background-color:rgb(255, 11, 11);
+        border-width:10px;
+    }
+   
+}
