+++ bytecodelearners-vue-js/src/views/winners/event1/css/faq.css
@@ -1,0 +1,3 @@
+#title{font-size: 40px;font-style: italic;font-weight: bold;
+    text-align: center;background-color: black;color: white;}
+#list{font: size 35px;font-weight: bold;}
