+++ bytecodelearners-vue-js/src/views/winners/event1/css/form.css
@@ -1,0 +1,26 @@
+div {
+  font-size: 100px;
+}
+
+.title {
+  font-size: 50px;
+  font-weight: bold;
+  font-style: italic;
+  background-color: black;
+  color: white;
+  text-align: center;
+}
+.form {
+  font-size: 25px;
+  font-weight: bold;
+}
+.submit {
+  text-align: center;
+}
+
+@media only screen and (max-width:400px) and(min-width:200){
+    .placeholder{
+        size: 8rem;
+    }
+
+}
