+++ bytecodelearners-vue-js/src/views/winners/event1/css/header.css
@@ -1,0 +1,64 @@
+html{
+    font-size: 62%;
+}
+
+
+body{
+    
+    background-color: rgba(145, 255, 249, 0.473);
+    background-size: 100%;
+
+}
+.main{
+    width: 10rem;
+    height: 2rem;
+    color: rgba(104, 26, 63, 0.842);
+}
+.contact{
+
+    float: left;
+    margin-left: .1rem;
+    margin-top: .4rem;
+}
+p{
+    font-size: 3rem;
+    color: rgb(22, 8, 8);
+}
+/* .view{
+    width: 1rem;
+    height: .6rem;
+    float: right;
+    margin-right: .4rem;
+    margin-top: .3rem;
+    border: .2rem solid black;
+     */
+/* } */
+nav{
+    font-size: 3rem;
+}
+div{
+    font-weight: bolder;
+    font-size: 7rem;
+}
+h1{
+    font-size:5rem;
+    color: rgb(105, 4, 67);
+}
+a{
+    font-size: 5rem;
+}
+h4{
+    font-size: 4rem;
+    color: rgba(112, 27, 27, 0.842);
+}
+@media(max-width:998px){
+    html {
+        font-size: 3rem;
+
+    }
+}@media(max-width:768px) {
+    html{
+        font-size: 3rem;
+    }
+}
+
