+++ bytecodelearners-vue-js/src/views/winners/event1/css/main.css
@@ -1,0 +1,86 @@
+body {
+  background-color: rgb(255, 255, 255);
+}
+ 
+  #box1 {
+    height: 250px;
+    width: 100%;
+    border: 2px solid rgb(68, 26, 26);
+    background-color: rgba(88, 20, 42, 0.404);
+    margin-top: 800px;
+    font-style: italic;
+    font-weight: bold;
+    padding-bottom: 10px;
+  }
+  #info li {
+    margin-left: 5px;
+  }
+  #info {
+    list-style: none;
+    font-size: 35px;
+  }
+   
+  #us{
+    font-size: 40px;
+  }
+
+  #box2 {
+    height: 200px;
+    width: 100%;
+    border: 1px solid rgb(68, 21, 21);
+    background-color: rgb(243, 241, 239);
+    text-align: center;
+  }
+  #box li {
+    background-color: rgb(9, 2, 15);
+  }
+  #box {
+    list-style: none;
+  }
+
+
+  
+  #box3 {
+    height: 250px;
+    width: 100%;
+    border: 1px solid rgb(70, 25, 25);
+    font-size: 2.5rem ;
+    font-weight: bold;
+    font-family: "Gill Sans", "Gill Sans MT", Calibri, "Trebuchet MS",
+      sans-serif;
+    background-color: rgba(240, 194, 194, 0.404);
+  }
+
+
+
+ 
+
+
+#box4 {
+  height: 250px;
+  width: 100%;
+  border: 1px solid rgb(77, 34, 34);
+  text-align: center;
+  background-color: rgba(78, 77, 78, 0.856);
+  font-size: 2.4rem;
+  margin-top: 1px;
+  /* padding: 10px;
+  padding-left: 30px;
+  padding-right: 40px; */
+  color: white;
+}
+
+@media screen  and (max-width:600px){
+  b{font-size:15px;}
+  #box1{font-size: 18px;width: 100%;height: 25%;}
+  #us{font-size: 30px;}
+  #box2{font-size: 20px; width: 100%;height:25% ;}
+  #info {font-size: 24px;}
+
+
+
+  #box3{font-size: 15px;width: 100%;height: 25%;}
+  #box4{font-size: 20px;width: 100%;height: 25%;}
+  
+}
+
