+++ bytecodelearners-vue-js/src/views/winners/event1/css/motion.css
@@ -1,0 +1,71 @@
+html{
+    font-size: 62%;
+}
+
+
+@import url('https://fonts.googleapis.com/css2?family=Alata&display=swap');
+
+ *{
+   font-family: 'Alata', sans-serif;
+   font-size: 4rem;
+}
+
+body{
+    background: rgba(238, 82, 100, 0.233);
+}
+
+h2
+{
+    position: relative;
+    font-size: 10rem;
+    color: rgba(255, 163, 163, 0.945);
+    -webkit-text-stroke:0.3rem rgb(82, 10, 10) ;
+   
+} 
+
+h2::before{
+    content:"ADHYAYAN...!!";
+    position: absolute;
+    color: rgb(107, 26, 26);
+    overflow: hidden;
+    animation: animate 2s linear infinite;
+}
+
+@keyframes animate {
+   from {
+        width: 0%;
+    }
+    
+    to {
+        width: 100%;
+    }
+}
+       
+h3{
+    font-size: 3rem;
+}
+p{
+    font-family: cursive;
+}
+
+@media(max-width:998px){
+    html {
+        font-size: 55%;
+
+    }
+}@media(max-width:768px) {
+    html{
+        font-size: 45%;
+    }
+
+    h2{
+        font-size: 7rem;
+    }
+    h3{
+        font-size: 5rem;
+    }
+    p{
+        font-size: 3rem;
+        color: rgb(190, 42, 42);
+    }
+}
