+++ bytecodelearners-vue-js/src/views/winners/event1/css/status.css
@@ -1,0 +1,5 @@
+.title{text-align: center;font-size: 35px;font-weight: bold;font-style: italic;
+    margin-top: 25px;   color: brown;}
+ .list{text-align: center; font-size: 20px;
+        font-weight: bold;}
+
