+++ bytecodelearners-vue-js/src/views/winners/event1/css/team.css
@@ -1,0 +1,13 @@
+.group{ background-color: rgba(140, 231, 247, 0.349);  height: 800px;width: 100%;}
+.title{text-align: center;font-size: 90px;font-style: italic;font-weight: bold;
+    padding-top: 70px;  color:black;  }
+    @media screen and (max-width:600px ) {.title{font-size: 40px;}}
+.team{font-size: 35px;float: left;margin: 13px;margin-top: 150px;
+text-align: center;    border: 3px solid white;}
+@media screen and (max-width:600px ) {.team{font-size: 15px;margin: 2px;margin-top: 50px;}}
+#mem1{background-color:bisque;height: 400px;width: 30%;}
+@media screen and (max-width:600px ) {#mem1{height: 250px;width: 30%;}}
+#mem2{background-color: bisque;height: 400px;width: 30%;}
+@media screen and (max-width:600px ) {#mem2{height: 250px;width: 30%;}}
+#mem3{background-color:bisque ;width: 30%;height: 400px;}
+@media screen and (max-width:600px ){#mem3{height: 250px;width: 30%;}}
