+++ bytecodelearners-vue-js/src/views/winners/event1/css/web.css
@@ -1,0 +1,103 @@
+html {
+  font-size: 62%;
+}
+
+/* .body{
+    background-color: rgb(245, 244, 244); */
+/* background-image:url(background-3.jpeg);
+    background-size: 100% ;
+} */
+
+/* .logobox {
+  width: 134rem;
+  height: 9rem;
+
+  border: 0.2rem solid rgb(90, 27, 27);
+} */
+
+#header {
+  background-color: rgba(247, 137, 139, 0.479);
+  height: 5rem;
+  text-align: center;
+  margin-top: 2.8rem;
+  margin-bottom: -30px;
+}
+a:hover {
+  color: rgb(83, 3, 3);
+  font-size: 3rem;
+}
+/* transform property and scale value */
+/* 
+.boxa {
+    background-color:rgb(250, 168, 168);
+    width: 135rem;
+    height: 4rem;
+    margin-top:3rem;
+    margin-bottom: 1.5rem; */
+    /* opacity: 0.6; */
+
+
+/* 
+.boxb{
+    background-color: rgb(253, 250, 250);
+    width: 135rem;
+    height: 22rem; */
+
+
+
+/* @media screen and (max-width:600px ) {.box4{font-siz?ter;} */
+/* 
+.boxc {
+    margin-top: 5rem;
+    background-color: rgba(68, 6, 6, 0.89);
+    width: 135rem;
+    height: 30rem; */
+
+
+/* .boxd{ */
+    /* background-color: rgba(140, 252, 252, 0.603);
+    width: 135rem;
+    height: 3rem;
+    font-size:x-large; */
+    /* font-style: oblique; */
+        /* text-align: center; } */
+
+
+h1{
+    
+    color: hsl(341, 71%, 19%);
+    font-size: 4rem;
+    margin: top .2rem;
+}  
+
+a {
+  color: rgba(80, 2, 9, 0.986);
+  font-weight: bolder;
+  text-decoration: none;
+  font-size: 2rem;
+}
+
+nav a {
+  font-size: 2rem;
+  padding-left: 25px;
+  padding-right: 1rem;
+  color: rgb(14, 1, 36);
+  font-weight: bold;
+}
+h2 {
+  font-size: 50px;
+  font-style: inherit;
+  color: hsl(341, 71%, 19%);
+}
+ p{
+    color: rgb(32, 1, 6);
+    font-size: 2rem;
+} 
+h3{
+    font-size: 3rem;
+    background-color: rgba(155, 211, 247, 0.39);
+}
+ h4,h5,h6{
+    color: rgb(32, 16, 16);
+     font-size: 2rem;
+}
