+++ bytecodelearners-vue-js/src/views/winners/event1/html/2@boxes.html
@@ -1,0 +1,46 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Box</title>
+<link rel="stylesheet" href="../css/boxstyl.css" type="text/css" media="screen"><link/>
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+    </head>
+  <body>
+    <div class="div1" id="a">
+      <h1>EXAM</h1>
+      <p>The candidate who are going to appear for various board exams ,university and entrance exams which are going to be held in 2021.</p>
+
+      <button style="font-size: 24px">
+        READ MORE<i class="fas fa-book-open"></i>
+      </button>
+    </div>
+
+    <div class="div2" id="b">
+      <h1>LOAN</h1>
+
+      <p>Getting an education loan is an easy way to finance your dreams . A student loan can help you get into the university of your choice.</p>
+
+      <button style="font-size: 24px">
+        READ MORE<i class="fas fa-book-open"></i>
+      </button>
+    </div>
+
+    <div class="div3" id="c">
+      <h1>COLLEGE</h1>
+      <p>Choose a collage to see it's rates and related information along with fee structure, campus life,environemnt, placement.</p>
+  
+
+      <button style="font-size: 24px">
+        READ MORE <i class="fas fa-book-open"></i>
+      </button>
+    </div>
+    <div class="div4" id="d">
+      <h1>SCHOLARSHIP</h1>
+      <p> Get the latest scholarship result 2021 and scholarship status here.Register for best scholarship according to your eligibility.   </p> <br />
+      <br />
+      <button style="font-size: 24px">
+        READ MORE<i class="fas fa-book-open"></i>
+      </button>
+    </div>
+  </body>
+</html>
