+++ bytecodelearners-vue-js/src/views/winners/event1/html/Adyayan about Us.html
@@ -1,0 +1,37 @@
+<html>
+<head>
+  
+    
+<title>About Us</title>
+<color-tex></color-tex>
+<link rel="stylesheet" href="../css/aboutus.css">
+</style>
+</head>
+<body>
+  <div class="header">
+<div class="main">
+  
+    <h1>ABOUT US</h1>
+   
+    <p>
+     <h2>Revolutionizing Education</h2> 
+      <br>
+      
+     Adhyayan are designed to give your career 
+     the perfect boost to
+      achive your full potential.
+      Empowring the student with access to 
+      better future.
+      <br>
+      <h4>Find every college,exam,scholarship and lone at one place</h4>
+
+
+    </p>
+   <button class="btn">Read More <i class="fas fa-book-open"></i></button>
+    
+
+
+</div>
+ </div>
+</body>
+</html>
