+++ bytecodelearners-vue-js/src/views/winners/event1/html/Boxes.html
@@ -1,0 +1,88 @@
+<!DOCTYPE html> <html> <head> <title>Box</title> 
+<style>
+
+
+@media only screen and(max-width:100px;)
+
+
+h1 {text-align: center;}
+p {text-align: center;}
+div {text-align: center;}
+
+
+ .div1{ height:300px; width:300px; background-color:teal; margin:0; }
+
+
+
+
+
+
+
+ .div2{ height:300px; width:300px; background-color:teal; position:relative; left:320px; bottom:300px; } 
+ 
+ 
+ .div3{ height:300px; width:300px; background-color:teal; position:relative; bottom:280px; }
+ 
+ 
+ 
+  .div4{ height:300px; width:300px; background-color:teal; position:relative; left:320px; bottom:580px; } 
+  
+  .div1 {
+  word-wrap: break-word;
+}
+.    div2 {
+  word-wrap: break-word;
+}
+    .div3 {
+  word-wrap: break-word;
+}
+    .div4{
+  word-wrap: break-word;
+}
+  
+  
+  </style> 
+  </head> 
+  <body> 
+  
+  <div class="div1"><h1>EXAM</h1> 
+   <p>vsywjeyshegeyeywhwywhuwu3gyqhhwhhwgehwgegeheheyehehehdbhdbddbbdh</p>
+   <p>Used on a button:</p>
+<button style='font-size:24px'>READ <i class='fas fa-book-open'></i></button>
+
+      
+  </div> 
+  
+ 
+  
+  <div class="div2"><h1>LONE</h1>  
+  
+  <p>vsywjeyshegeyeywhwywfgh
+  <br>cuufducufuffuccj<br>fsjgxhkdoydykxg</p>
+  
+   <p>Used on a button:</p>
+<button style='font-size:24px'>READ <i class='fas fa-book-open'></i></button>
+
+      
+      
+      
+  </div> 
+  
+  
+  <div class="div3"><h1>COLLEGE</h1>
+  <p>vsywjeyshegeyeywhwywhuwu3gyqhhwhhwg>ehwgegeheheyehehehdbhdbddbbdh</p> 
+  
+   <p>Used on a button:</p>
+<button style='font-size:24px'>READ <i class='fas fa-book-open'></i></button>
+
+   </div>
+   <div class="div4"><h1>SCHOLARSHIP</h1>
+       <p>vsywjeyshegeyeywhwywhuwu3gyqhhwhhwgehwgegeheheyehehehdbhdbddbbdh</p>
+        <p>Used on a button:</p>
+<button style='font-size:24px'>READ <i class='fas fa-book-open'></i></button>
+       
+   </div>
+   
+    </body> 
+    </html>
+
