+++ bytecodelearners-vue-js/src/views/winners/event1/html/anime.html
@@ -1,0 +1,82 @@
+ <html> <head> <title>Box</title> 
+    <style>
+    
+    <div class = "row">
+      <input type="button" style="background-color:black;color:white;width:150px;
+      height:40px;>
+    
+    
+    h1 {text-align: center;}
+    p {text-align: center;}
+    div {text-align: center;}
+    
+    
+     .div1{ height:300px; width:300px; background-color:rgba(156, 245, 245, 0.534); margin:0; }
+    
+ .div2{ height:300px; width:300px; background-color:rgb(153, 245, 245); position:relative; left:320px; bottom:300px; } 
+ 
+ 
+    .div3{ height:300px; width:300px; background-color:rgb(176, 245, 245); position:relative; bottom:280px; f }
+ 
+ 
+ 
+  .div4{ height:300px; width:300px; background-color:rgb(163, 241, 241); position: relative; ; left:320px; bottom:580px; } 
+  
+  .div1 {
+  word-wrap: break-word;
+}.div1 {
+  word-wrap: break-word;
+}
+  .div2 {
+  word-wrap: break-word;
+}
+    .div3 {
+  word-wrap: break-word;
+}
+    .div4{
+  word-wrap: break-word;
+}</style> 
+  </head> 
+  <body>
+
+  
+<div class="div1"><h1>EXAM</h1> 
+   <p Course Details>Updates, Notification,Admit cards,Result,Books,Syllabus and exams.</p>
+   
+<button style='font-size:24px'>READ <i class='fas fa-book-open'></i></button>
+
+      
+  </div> 
+  
+ 
+  <div class="div2"><h1>LOANS</h1>  
+    <p>Apply for loans,bridge the gap in education finance.Smart Loans To Accelerate Your Career. </p>
+   
+    <button style='font-size:24px'>READ <i class='fas fa-book-open'></i></button>
+    
+          
+          
+          
+      </div> 
+       
+  <div class="div3"><h1>COLLEGE</h1>
+    <p>GO Confidentely in the DIRECTION of your Dreams!!.Live the LIFE you've imagined</p> 
+    
+     
+  <button style='font-size:24px'>READ <i class='fas fa-book-open'></i></button>
+  
+     </div>
+     <div class="div4"><h1>SCHOLARSHIP</h1>
+         <p><A>Apply for scholarship , find the best scholariship as per your eligiblity.Turning Educational dreams into reality</A></p>
+          <br>
+          <br>
+          <button style='font-size:24px'>READ <i class='fas fa-book-open'></i></button>
+       
+        </div>
+        
+         </body> 
+         </html>
+     
+  
+  
+
