+++ bytecodelearners-vue-js/src/views/winners/event1/html/box.html
@@ -1,0 +1,47 @@
+<html>
+<head>
+<title>MIDDLE PART</title>
+</head>
+<link rel="stylesheet" type="text/css" media="screen" href="../css/box.css"><link/>
+<hr style="background-color: black;height: 3px;">
+<div class="title">WINGS YOUR IMAGINATION</div><hr style="height: 3px;background-color: black;">
+<hr style="background-color: black;height: 6px;">
+ 
+<div class="box1" id="a"><ul style="list-style: none;color:black;
+font-size: 35px;margin-left: -50px;background-color: white;">
+<li>EXAM</li></ul><hr style="height: 2px;
+background-color: white;margin-top: -6px;">
+<p>EXAM UPDATES,<br>NOTIFICATIONS,<br>ADMIT CARDS,RESULTS,
+<br>BOOKS,SYLLABUS and COURSE DETAILS</p>
+<p><button style="background-color:white;color:black;font-size: 20px;
+font-weight: bold;border-color: white;">READ MORE</button></p></div>
+
+
+<div class="box1" id="b"><ul style="list-style: none;color:black;
+    font-size: 35px;margin-left: -50px;background-color: white;">
+    <li>LOAN</li></ul><hr style="height: 2px;
+    background-color: white;margin-top: -6px;">
+    <p>APPLY FOR LOANS,<br>BRIDGE THE GAP IN EDUCATION FINANACE,<br>SMART LOANS TO ACCELERATE CAREER</p>
+    <p><button style="background-color:white;color:black;font-size: 20px;
+    font-weight: bold;border-color: white;">READ MORE</button></p></div>
+
+
+<div class="box1" id="c"><ul style="list-style: none;color:black;
+    font-size: 35px;margin-left: -50px;background-color: white;">
+    <li>COLLEGE</li></ul><hr style="height: 2px;
+    background-color: white;margin-top: -6px;">
+    <p>GO CONFIDENTLY IN THE DIRECTION of YOUR DREAM! LIVE THE LIFE YOU HAVE IMAGINED</p>
+    <p><button style="background-color:white;color:black;font-size: 20px;
+    font-weight: bold;border-color: white;margin-bottom: -20px;">READ MORE</button></p></div>
+
+
+<div class="box1" id="d"><ul style="list-style: none;color:black;
+    font-size: 35px;margin-left: -50px;background-color: white;">
+    <li>SCHOLARSHIP</li></ul><hr style="height: 2px;
+    background-color: white;margin-top: -6px;">
+    <p>FIND THE BEST SCHOLARSHIP AS PER ELIGIBILITY,TURNING EDUCATIONAL DREAMS IN TO REALITY</p>
+    <p><button style="background-color:white;color:black;font-size: 20px;
+    font-weight: bold;border-color: white;">READ MORE</button></p></div>
+  
+</body>
+</html>
