+++ bytecodelearners-vue-js/src/views/winners/event1/html/faq.html
@@ -1,0 +1,23 @@
+<html>
+<head>
+<title>FAQ's</title>
+<link rel="stylesheet" type="text/css" media="screen" href="../css/faq.css"><link/>
+</head>
+<body>
+<div class="faq" id="title">Get Answer Of All Your Questions
+ About Scholarship</div>
+<div class="faq" id="list"><hr style="margin-top: 8px;height: 4px;background-color: black;"><ul> <li style="margin-top: 40px;">
+SCHOLARSHIP FOR CLASS 10th AND 12th PASSED STUDENTS</li>
+<li>SCHOLARSHIP FOR GRADUATE STUDENTS</li>
+<li>EVERYTHING ABOUT EDUCATION LOANS IN INDIA</li>
+<li>CBSE or ICSE</li>
+<li>COURSES IN MEDICAL FIELD OTHER THAN MBBS</li>
+<li>STUDY ABROAD-FINANCE:A DECIDING FACTOR</li>
+<li>TOP 20 ENGINEERING COLLEGES OF UPTU</li>
+<li>COURSES AVAILABLE AFTER CLASS 10th</li>
+<li>LIST THE TOP GOVT COLLEGE UNDER UPTU</li>
+<li>COMPETITIVE EXAMS FOR FURTHER STUDIES AFTER 12th</li>
+</ul>
+</div>
+</body>
+</html>
