+++ bytecodelearners-vue-js/src/views/winners/event1/html/form.html
@@ -1,0 +1,34 @@
+<html>
+<head><title>SCHOLARSHIP FORM</title>
+<link rel="stylesheet" type="text/css" media="screen" href="../css/form.css"><link/>
+<link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
+    <link/>
+</head>
+<body>
+<div class="title">LOGIN / REGISTER</div>
+<div class="form">
+<ul style="margin-top: 35px;"><li>STUDENT's PHOTO : </li>
+ <li style="list-style:none;"><div class="photo" style="width: 14%;height: 25%;border: 1px solid black;">
+<i class="material-icons" style="font-size: 150px;text-align: center;"></i></div></li><br>
+<li>STUDENT's NAME : <input placeholder="LAST FIRST MIDDLE"></li>
+<br><li>FATHER's NAME : <input placeholder=""></li><br>
+<li>MOTHER's NAME : <input placeholder=""></li><br>
+<li>DATE of BIRTH : <input placeholder="MONTH-DAY-YEAR"></li><br>
+<li>STATE of ORIGIN: <input placeholder=""></li><br>
+<li>HOME ADDRESS : <input placeholder=""></li><br>
+<li>PHONE NO. : <input placeholder=""></li><br>
+<li>EMAIL ID : <input placeholder=""></li><br>
+<li>AADHAR NO. : <input placeholder=""></li><br>
+<li>AGE : <input placeholder=""></li><br>
+<li>GENDER : <input placeholder=""></li><br>
+<li>10th BOARD : <input placeholder=""></li><br>
+<li>10th SCORE : <input placeholder=""></li><br>
+<li>12th BOARD : <input placeholder=""></li><br>
+<li>12th SCORE : <input placeholder=""></li><br>
+<li>OTHER COMPETITIVE EXAM / SCORE : <input placeholder=""></li><br>
+<li>ANNUAL SALARY : <input placeholder=""></li><br>
+<li>WHY DO YOU NEED THIS SCHOLARSHIP :</li><li style="list-style: none;"> <input placeholder="max 5 lines"
+     style="width: 100%;height: 25%;"></li><br>
+</ul>   
+</div>
+<div class="submit" ><button>SUBMIT</button></div></body></html>
