+++ bytecodelearners-vue-js/src/views/winners/event1/html/header.html
@@ -1,0 +1,35 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <title></title>
+        <link rel="stylesheet" href="../css/header.css" type="text/css">
+    </head>
+    <body>
+        <div class="main">
+            <div class="contact">Follow</div><br>
+           <div class="view"> </div>
+        </div>
+
+       <br><br><br><br><br><br> <h2> CONNECT ON LINKDIN :)</h2>
+
+      <nav>
+        <a href="https://www.linkedin.com/in/susmita-sagar-706179209" target="_blank">Linkdin Profile Link</a>
+        <h1></h1>
+      </nav>
+
+
+     <h3> CONNECT ON FACEBOOK :)</h3>
+      
+        <a href="https://www.facebook.com/susmita.sagar.77"target="_blank">Facebook Link</a>
+       `  
+       <h4 > Contact Us Through Email id :-</h4> 
+
+       <P>susmitasagar038@gmail.com</P>
+       <p>umeshkumari123029@gmail.com</p>
+       <p>archu0943@gmail.com</p>
+
+      </div>
+      
+
+    </body>
+</html>
