+++ bytecodelearners-vue-js/src/views/winners/event1/html/index1.html
@@ -1,0 +1,61 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset='utf-8'>
+    <meta http-equiv='X-UA-Compatible' content='IE=edge'>
+    <title>ADHAYAPAN</title>
+    <meta name='viewport' content='width=device-width, initial-scale=1'>
+    <link rel='stylesheet' type='text/css' media='screen' href='../css/main.css'><link/>
+    <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
+    <link/>
+    <script src='../js/'></script>
+</head>
+<body>
+<div class="container" id="box1">
+    <b style="font-size: 50px; color: white;">CONTACT US:-</b>
+<ul id="info"><li style="color: white;"> <i class="material-icons" style="font-size: 35px;" >phone</i>9996547883 (Mon-Fri,10am to 6pm) </li>
+
+<li style="color: white;"><i class="material-icons" style="font-size: 35px;"> mail</i>cuhmahendergarh@gmail.com</li>
+</ul>
+</div>
+<div class="container" id="box2"><ul id="box" style="color: white;padding: 10px;text-align: center;" >
+    <button style="background-color: black;color: white;font-size: 35px;"><b>ABOUT US</b></button>
+  <button style=" color :white;background-color: black;font-size: 35px;"><b>TEAM</b></button>  
+
+    
+<button style="  background-color:black; color:white;font-size: 35px;"><b>SCHOLARSHIP RESULT</b> 
+
+</button>
+<button style="  background-color:black; color:white;font-size: 35px;"> <b>MEDIA PARTNERS</b>
+
+</button>
+<button style="  background-color:black; color:white;font-size: 35px;" ><b>FAQ's</b>
+
+</button>
+<button style="  background-color:black; color:white;font-size: 35px;" ><b>CHECK YOUR STATUS</b></button>
+ </ul> 
+</div>
+<div class="container" id="box3"><marquee behavior="scroll" direction="up"><ol>
+<li>CURRENT CLASS SCHOLARSHIP</li>
+<li>STATEWISE SCHOLARSHIP</li>
+<li>SCHOLARSHIP FOR STUDY ABROAD</li>
+<li>STUDY LOANS</li>
+<li>SCHOLARSHIP FOR ENGINEERING/MEDICAL STUDENTS</li>
+</ol><button style="float: left; margin-left: 8px;font-weight: bold;">more</button>
+</marquee></div>
+<div  class="container" id="box4" style="color: rgb(252, 252, 247);" >
+<b>TERMS & CONDITIONS</b><b>|</b><b>PRIVACY POLICES</b><b>|</b>
+<b>DISCLAIMER</b><b>|</b><b>CAREERS</b><b>|</b>
+<ul id="logo">
+<li><a href="https://www.twitter.com" >twitter</a></li>
+<li><a href="https://www.google.com" >google</a></li>
+<li><a href="https://www.facebook.com" >facebook</a></li>
+<li><a href="https://www.instagram.com">instagram</a></li>
+<li><a href="https://www.linkedin.com">linkedin</a></li>
+</ul>
+</div>
+
+
+
+</body>
+</html>
