+++ bytecodelearners-vue-js/src/views/winners/event1/html/logo.html
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>logo</title>
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css" integrity="sha256-eZrrJcwDc/3uDhsdt61sL2oOBY362qM3lon1gyExkL0=" crossorigin="anonymous" />
+    <link rel="stylesheet" href="../css/logo.css"/>
+
+</style>
+
+</head>
+<body>
+    <div class="centerdiv">
+        <a href="#">
+          <i class="fa fa-3x fa-facebook"> </i>
+        </a>
+
+        <a href="#">
+            <i class="fa fa-3x fa-instagram"></i>
+        </a>
+
+        <a href="#">
+            <i class="fa fa-3x fa-twitter"></i>
+        </a>  
+        
+        <a href="#">
+            <i class="fa fa-3x fa-linkedin"></i>
+        </a>
+        
+        <a href="#">
+            <i class="fa fa-3x fa-youtube"></i>
+        </a>
+    </div>
+</body>
+</html>
