+++ bytecodelearners-vue-js/src/views/winners/event1/html/menu1.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Menu</title>
+    <link rel="stylesheet" href="../css/menu1.css">
+</head>
+<body>
+
+    <center><img class="pic1" src="../image/adhyayan logo.jpg"</center>
+   <h1>-: MENU :-</h1>
+    <p>This website includes many important updates regarding a student's carrier</p>
+    <p>We provide following information :-</p>
+    <h2># SCHOLARSHIPS</h2>
+    <h3># EXAM</h3>
+    <h4># FOREIGN STUDIES</h4>
+    <h5># LOANS</h5>
+    <h6># COLLAGES</h6>
+    <a href="../html/index.html">BACK</a>
+
+    
+    
+    <div>
+        
+    </div>
+
+
+</body>
+</html>
