+++ bytecodelearners-vue-js/src/views/winners/event1/html/motion.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Motion</title>
+    <link rel="stylesheet" href="../css/motion.css">
+</head>
+    <h2> ADHYAYAN...!! </h2>
+      <h3> wings to carrier..!! </h3>
+      <p>  ::Updates, Notification,Admit cards,Result,Books,Syllabus and Exams </p>
+      <p>  :: Apply for loans,bridge the gap in education finance.Smart Loans To Accelerate Your Career. </p>
+    
+</html>
+</body>
