+++ bytecodelearners-vue-js/src/views/winners/event1/html/result.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset='utf-8'>
+    <meta http-equiv='X-UA-Compatible' content='IE=edge'>
+    <title>SCHOLARSHIP RESULT</title>
+    <meta name='viewport' content='width=device-width, initial-scale=1'>
+    <link rel='stylesheet' type='text/css' media='screen' href='../css/result.css'>
+    <script src='../js/'></script>
+</head>
+<body>
+<div class="title">SCHOLARSHIP RESULT</div>
+<hr style="line-height: 4px;color: brown;">
+<hr style="line-height: 4px;color: brown;">
+<div class="list" id="1" style="margin-top: 100px;">APPLICATION ID</div>
+<ul class="list" id="3" style="list-style: none;"><li><input placeholder=""></li></ul>
+<div class="list" id="2" >PASSWORD</div>
+<ul class="list" id="4" style="list-style: none;"><li><input placeholder=""></li></ul>
+<div class="list">
+<button  >CHECK</button></div><hr style="line-height: 4px;color: brown;margin-top: 160px;">
+<hr style="line-height: 4px;color: brown;">
+</body>
+</html>
