+++ bytecodelearners-vue-js/src/views/winners/event1/html/slider.html
@@ -1,0 +1,16 @@
+<html>
+<head>
+<title>PHOTOS </title>
+<link rel='stylesheet' type='text/css' media='screen' href='slider.css'>
+</head>
+<body><marquee behaviour="scroll" direction="left" scrollamount="20">
+<div  class="photo"  id="a"><img src="../image/uni.jpg" style="height: 100%;width: 100%;">
+    <img src="../image/bank-education-loan[66].jpg" style="height: 100%;width: 100%;">
+<img src="../image/exam pic (1).jpeg" style="height: 100%;width:100% ;" >   
+<img src="../image/Courses For Education Loan[67].jpg" style="height: 100%;width: 100%;">   
+ <img src="../image/group.jpg" style="height: 100%;width: 100%;"></marquee> 
+ 
+</div>
+
+</body>
+</html>
