+++ bytecodelearners-vue-js/src/views/winners/event1/html/team.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset='utf-8'>
+    <meta http-equiv='X-UA-Compatible' content='IE=edge'>
+    <title>TEAM SECTION</title>
+    <meta name='viewport' content='width=device-width, initial-scale=1'>
+    <link rel='stylesheet' type='text/css' media='screen' href='../css/team.css'>
+</head>
+<body><div class="group"><div class="title" >TEAM MEMBER</div><hr><hr>
+<marquee behaviour="horizontal" direction="left">
+    <div class="team" id="mem1"><img src="../image/susmita.jpg" 
+style="height:80%;width: 100%;border-radius: 50%;" ><br>SUSMITA SAGAR </div>
+<div class="team" id="mem2"><img src="../image/umesh.jpg" 
+ style="height: 80%;width: 100%;border-radius: 50%;"><br>UMESH KUMARI</div>
+<div class="team" id="mem3"><img src="../image/archana.jpg" 
+style="height: 80%;width: 100%;border-radius: 50%;"><br>ARCHANA KUMARI</div></marquee>
+</div>
+</body>
+</html>
