+++ bytecodelearners-vue-js/src/views/winners/event1/html/web2.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>web2</title>
+    <link rel="stylesheet" href="../css/web2.css">
+</head>
+<body>
+
+</body>
+<div class="container">
+   <center> <img src="../image/adhyayan logo.jpg" alt="center"></center>
+
+</div>
+<h1> HOME :-
+    <p>Apply for exams,scholarships,collage,educatiom Loans with easy to fill form - most of your information will be prefilled and
+        Upload high-quality documents with all fields visible
+        Regularly monitor your email/mobile for Adhyayan notifications
+        Respond to questions in telephonic interviews
+    </p>
+</h1>
+
+
+<h2>Become a Scholar:-
+   <p> If you are selected as a scholar, you will receive notifications from Adhyayan
+    You can also check results on results page
+    Upload your attested documents on Adhyayan and provide bank information
+    Receive scholarship award directly in your bank account </p>
+</h2>
+ <center><a href="../html/index.html" target="_blank">Back</a></center>
+</html>
